requirements.txt
discord.py==0.16.12
requests==2.20.0